+++ changelog.md
@@ -1,2 +1,3 @@
 * Updated to 1.21.11
+* Added Vibranium, Adamantium & Nether Spears
 * Migrated code to Mojang mappings